--- Sources/Compass+Navigate.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-extension Compass {
-
-  public static func navigate(urn: String, scheme: String = Compass.scheme) {
-    guard let url = NSURL(string: "\(scheme)\(urn)") else { return }
-    openURL(url)
-  }
-}
